server/requirements.txt
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
psycopg2==2.9.1
pycparser==2.20
pyjwt==2.1.0
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2021.1
social-auth-core==4.1.0
sqlparse==0.4.0
